# docs/eai-eventportal-api.swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Portal API",
        "description": "API for Event Portal",
        "contact": {
            "email": "rod.pattison@woodside.com.au"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://events.api.woodside",
            "description": "Prod Events Platform API"
        },
        {
            "url": "https://events-test.dev.api.woodside",
            "description": "Non-prod Events Platform API"
        }
    ],
...
    "paths": {
        "/schemas": {
            "get": {
                "tags": ["schemas"],
                "summary": "fetches all schemas",
                "description": "Fetches all schemas from the event registry",
                "operationId": "fetchSchemas",
                "responses": {
                    "200": {
                        "description": "list of available schemas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SchemaSummaryDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["schemas"],
                "summary": "creates a new schema",
                "description": "Adds a schema to the registry.",
                "operationId": "addSchema",
                "requestBody": {
                    "description": "Schema item to add",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSchemaDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Schema created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchemaSummaryDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                }
            }
        },
        "/schemas/{schemaName}": {
            "get": {
                "tags": ["schemas"],
                "summary": "fetches a schema by name",
                "description": "Fetches a schema by the specified name from the event registry",
                "operationId": "fetchSchema",
                "responses": {
                    "200": {
                        "description": "the schema definition output",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/SchemaDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error getting schema: NotFoundException: Schema with name {schemaName} does not exist."
                    }
                }
            },
            "put": {
                "tags": ["schemas"],
                "summary": "updates a schema",
                "description": "Updates a schema in the registry with a new version.",
                "operationId": "updateSchema",
                "requestBody": {
                    "description": "Schema item to add",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSchemaDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Schema successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchemaSummaryDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                }
            }
        },
        "/schemas/{schemaName}/recent-events": {
            "get": {
                "tags": ["schemas"],
                "summary": "fetches recent events for a given schema",
                "description": "Fetches up to 10 of the most recent events that have been logged for that schema.",
                "operationId": "getRecentEvents",
                "parameters": [
                    {
                        "in": "query",
                        "name": "hours",
                        "description": "Number of hours in the past relative to now for which the query should search. Note that the default value is 1 if not passed, and values greater than 168 will be rejected.",
                        "defaultValue": "1",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the schema versions output",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/RecentEvent"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/schemas/{schemaName}/versions": {
            "get": {
                "tags": ["schemas"],
                "summary": "fetches all versions in a schema",
                "description": "Fetches a schema versions by the specified name from the event registry",
                "operationId": "fetchSchemaVersions",
                "responses": {
                    "200": {
                        "description": "the schema versions output",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/SchemaVersions"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error getting schema: NotFoundException: Schema with name {schemaName} does not exist."
                    }
                }
            }
        },
        "/schemas/{schemaName}/versions/{versionNumber}": {
            "get": {
                "tags": ["schemas"],
                "summary": "fetches schema by name and version",
                "description": "Fetches a schema by the specified name and version from the event registry",
                "operationId": "fetchVersionedSchema",
                "responses": {
                    "200": {
                        "description": "the schema versions output",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/SchemaDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/schemas/{schemaName}/versions/{versionNumber}/publish": {
            "post": {
                "tags": ["schemas"],
                "summary": "publishes a schema",
                "description": "publishes a schema in the registry and deprecates previous version.",
                "operationId": "publishSchema",
                "requestBody": {
                    "description": "Schema to publish",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublishSchemaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request accepted for processing but not completed"
                    },
                    "400": {
                        "description": "SchemaName is empty or not matching with path parameters"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SchemaSummaryDto": {
                "type": "object",
                "properties": {
                    "LastUpdated": { "type": "string" },
                    "SchemaArn": { "type": "string" },
                    "SchemaName": { "type": "string" },
                    "Tags": {
                        "type": "object",
                        "properties": {
                            "sysDomain": { "type": "string" },
                            "sysVersionState": { "type": "string" },
                            "sysDecription": { "type": "string" },
                            "sysVersionMap": { "type": "string" }
                        },
                        "additionalProperties": true
                    },
                    "VersionCount": { "type": "integer" },
                    "CostCode": { "type": "string" },
                    "SchemaOwner": { "type": "string" },
                    "SchemaSupportGroup": { "type": "string" },
                    "EventClassification": { "type": "string" },
                    "AppInstId": { "type": "string" },
                    "OwnerRole": { "type": "string" },
                    "Description": { "type": "string" },
                    "AvailableVersions": {
                        "type": "object",
                        "additionalProperties": { "type": "string", "enum": ["publ", "drft", "depr"] }
                    }
                }
            },
            "SchemaDetails": {
                "type": "object",
                "properties": {
                    "Content": { "type": "string" },
                    "Description": { "type": "string" },
                    "LastUpdated": { "type": "string" },
                    "SchemaArn": { "type": "string" },
                    "SchemaName": { "type": "string" },
                    "SchemaVersion": { "type": "string" },
                    "AppInstId": { "type": "string" },
                    "OwnerRole": { "type": "string" },
                    "CostCode": { "type": "string" },
                    "SchemaOwner": { "type": "string" },
                    "SchemaSupportGroup": { "type": "string" },
                    "EventClassification": { "type": "string" },
                    "Tags": {
                        "type": "object",
                        "properties": {
                            "sysDomain": { "type": "string" },
                            "sysVersionState": { "type": "string" },
                            "sysDecription": { "type": "string" },
                            "sysVersionMap": { "type": "string" }
                        }
                    },
                    "Type": { "type": "string" },
                    "VersionCreatedDate": { "type": "string" },
                    "VersionCount": { "type": "integer" },
                    "AvailableVersions": {
                        "type": "object",
                        "additionalProperties": { "type": "string", "enum": ["publ", "drft", "depr"] }
                    }
                }
            },
            "SchemaVersions": {
                "type": "object",
                "additionalProperties": { "type": "string", "enum": ["publ", "drft", "depr"] }
            },
            "PublishSchemaRequest": {
                "type": "object",
                "properties": {
                    "SchemaName": { "type": "string" },
                    "PlatformVersion": { "type": "string" }
                }
            },
            "CreateSchemaDto": {
                "type": "object",
                "required": [
                    "Content",
                    "SchemaName",
                    "Tags",
                    "AppInstId",
                    "OwnerRole",
                    "CostCode",
                    "SchemaOwner",
                    "SchemaSupportGroup",
                    "EventClassification"
                ],
                "properties": {
                    "Description": { "type": "string" },
                    "Content": { "type": "string" },
                    "SchemaName": { "type": "string" },
                    "AppInstId": { "type": "string" },
                    "OwnerRole": { "type": "string" },
                    "CostCode": { "type": "string" },
                    "SchemaOwner": { "type": "string" },
                    "SchemaSupportGroup": { "type": "string" },
                    "EventClassification": { "type": "string" },
                    "Tags": {
                        "type": "object"
                    }
                }
            },
            "UpdateSchemaDto": {
                "type": "object",
                "required": [
                    "Content",
                    "SchemaName",
                    "Tags",
                    "AppInstId",
                    "OwnerRole",
                    "CostCode",
                    "SchemaOwner",
                    "SchemaSupportGroup",
                    "EventClassification"
                ],
                "properties": {
                    "Description": { "type": "string" },
                    "Content": { "type": "string" },
                    "SchemaName": { "type": "string" },
                    "AppInstId": { "type": "string" },
                    "OwnerRole": { "type": "string" },
                    "CostCode": { "type": "string" },
                    "SchemaOwner": { "type": "string" },
                    "SchemaSupportGroup": { "type": "string" },
                    "EventClassification": { "type": "string" },
                    "Tags": {
                        "type": "object"
                    }
                }
            },
...
            "RecentEvent": {
                "type": "array",
                "items": {
                    "type": "object",
                    "properties": {
                        "account": {
                            "type": "string"
                        },
                        "detail": {
                            "type": "object",
                            "properties": {
                                "Data": {
                                    "type": "string",
                                    "description": "User defined event shape, could be any type"
                                },
                                "Metadata": {
                                    "type": "object",
                                    "properties": {
                                        "BusinessKey": {
                                            "type": "string"
                                        },
                                        "Guid": {
                                            "type": "string"
                                        },
                                        "MaintenancePlant": {
                                            "type": "string"
                                        },
                                        "Origin": {
                                            "type": "string"
                                        },
                                        "TechnicalObjectType": {
                                            "type": "string"
                                        },
                                        "Time": {
                                            "type": "string"
                                        },
                                        "Version": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "BusinessKey",
                                        "Guid",
                                        "MaintenancePlant",
                                        "Origin",
                                        "TechnicalObjectType",
                                        "Time",
                                        "Version"
                                    ]
                                }
                            },
                            "required": ["Data", "Metadata"]
                        },
                        "detail-type": {
                            "type": "string"
                        },
                        "id": {
                            "type": "string"
                        },
                        "region": {
                            "type": "string"
                        },
                        "resources": {
                            "type": "array",
                            "items": {}
                        },
                        "source": {
                            "type": "string"
                        },
                        "time": {
                            "type": "string"
                        },
                        "version": {
                            "type": "string"
                        }
                    },
                    "required": [
                        "account",
                        "detail",
                        "detail-type",
                        "id",
                        "region",
                        "resources",
                        "source",
                        "time",
                        "version"
                    ]
                }
            }
        }
    }
}